LTS updated to 5.10.233

Release date: 2025-03-19
TAG: android12-5.10-2025-02_r3
Deprecation date: 2026-03-01

kernel Artifacts: https://ci.android.com/builds/submitted/13232202/kernel_aarch64/latest
source tag: https://android.googlesource.com/kernel/common/+/refs/tags/android12-5.10-2025-02_r3
release URL: https://source.android.com/docs/core/architecture/kernel/gki-android12-5_10-release-builds#february-2025-releases

+++ drivers/gpu/drm/msm/disp/dpu1/dpu_core_perf.c
@@ -79,7 +79,7 @@ static u64 _dpu_core_perf_calc_clk(struct dpu_kms *kms,
 
 	mode = &state->adjusted_mode;
 
-	crtc_clk = mode->vtotal * mode->hdisplay * drm_mode_vrefresh(mode);
+	crtc_clk = (u64)mode->vtotal * mode->hdisplay * drm_mode_vrefresh(mode);
 
 	drm_atomic_crtc_for_each_plane(plane, crtc) {
 		pstate = to_dpu_plane_state(plane->state);
